+++ includes/Hooks/RelatedArticlesHandler.php
@@ -80,16 +80,17 @@ class RelatedArticlesHandler implements
 	 */
 	private function getRelatedTitles( Title $title, $limit ): array {
 		$dbr = $this->loadBalancer->getConnectionRef( ILoadBalancer::DB_REPLICA );
+		$namespaces = (array)$this->config->get( 'UnifiedExtensionForFemiwikiRelatedArticlesTargetNamespaces' );
 
 		// Make a UNION query
-		// $select1, $select2,
 		$union = $dbr->unionQueries( [
-			$this->makeTitlesFromHereSQL( $title, $dbr ),
-			$this->makeRedirectedTitlesFromHereSQL( $title, $dbr ),
-			$this->makeTitlesToHereSQL( $title, $dbr ),
-			$this->makeTitlesToRedirectsOfHereSQL( $title, $dbr ),
+			$this->makeTitlesFromHereSQL( $title, $dbr, $namespaces ),
+			$this->makeRedirectedTitlesFromHereSQL( $title, $dbr, $namespaces ),
+			$this->makeTitlesToHereSQL( $title, $dbr, $namespaces ),
+			$this->makeTitlesToRedirectsOfHereSQL( $title, $dbr, $namespaces ),
 		], $dbr::UNION_DISTINCT ) .
-			' ORDER BY page_touched ' . SelectQueryBuilder::SORT_DESC;
+			// Give higher priority to main namespace.
+			' ORDER BY page_namespace, page_touched ' . SelectQueryBuilder::SORT_DESC;
 		$sql = $dbr->limitResult( $union, $limit );
 		// Do not use dbr->select() because the core does not have a method for UNION followed by ORDER BY.
 		// phpcs:ignore MediaWiki.Usage.DbrQueryUsage.DbrQueryFound
@@ -115,12 +116,14 @@ class RelatedArticlesHandler implements
 	 *
 	 * @param Title $title
 	 * @param DBConnRef $dbr
+	 * @param array $targetNamespaces Empty array means accepting all namespaces.
 	 * @return string
 	 */
-	private function makeTitlesFromHereSQL( Title $title, DBConnRef $dbr ): string {
+	private function makeTitlesFromHereSQL( Title $title, DBConnRef $dbr, $targetNamespaces ): string {
 		return $dbr->newSelectQueryBuilder()
 			->table( 'pagelinks' )
 			->conds( [ 'pl_from' => $title->getArticleId() ] )
+			->conds( $targetNamespaces ? [ 'pl_namespace' => $targetNamespaces ] : [] )
 			->leftJoin( 'page', 'page', [
 				'page_namespace = pl_namespace',
 				'page_title = pl_title',
@@ -145,12 +148,14 @@ class RelatedArticlesHandler implements
 	 *
 	 * @param Title $title
 	 * @param DBConnRef $dbr
+	 * @param array $targetNamespaces Empty array means accepting all namespaces.
 	 * @return string
 	 */
-	private function makeRedirectedTitlesFromHereSQL( Title $title, DBConnRef $dbr ): string {
+	private function makeRedirectedTitlesFromHereSQL( Title $title, DBConnRef $dbr, $targetNamespaces ): string {
 		return $dbr->newSelectQueryBuilder()
 			->table( 'pagelinks' )
 			->conds( [ 'pl_from' => $title->getArticleId() ] )
+			->conds( $targetNamespaces ? [ 'pl_namespace' => $targetNamespaces ] : [] )
 			->leftJoin( 'page', 'link_target', [
 				'link_target.page_namespace = pl_namespace',
 				'link_target.page_title = pl_title',
@@ -165,7 +170,7 @@ class RelatedArticlesHandler implements
 				'target.page_title = rd_title',
 			] )
 			->conds( [
-				// Hide redirects
+				// Only redirects
 				'rd_from != 0',
 			] )
 			->fields( [
@@ -181,21 +186,27 @@ class RelatedArticlesHandler implements
 	 *
 	 * @param Title $title
 	 * @param DBConnRef $dbr
+	 * @param array $targetNamespaces Empty array means accepting all namespaces.
 	 * @return string
 	 */
-	private function makeTitlesToHereSQL( Title $title, DBConnRef $dbr ): string {
+	private function makeTitlesToHereSQL( Title $title, DBConnRef $dbr, $targetNamespaces ): string {
 		return $dbr->newSelectQueryBuilder()
 			->table( 'pagelinks' )
+			->conds( [
+				'pl_namespace' => $title->getNamespace(),
+				'pl_title' => $title->getDBkey(),
+			] )
 			->leftJoin( 'redirect', 'redirect', [ 'rd_from = pl_from' ] )
 			->leftJoin( 'page', 'page', [
 				'page_id = pl_from'
 			] )
 			->conds( [
-				'pl_namespace' => $title->getNamespace(),
-				'pl_title' => $title->getDBkey(),
 				// Hide redirects
 				'rd_from' => null,
 			] )
+			->conds( $targetNamespaces ? [
+				'page_namespace' => $targetNamespaces,
+			] : [] )
 			->fields( [
 				'page_namespace',
 				'page_title',
@@ -209,9 +220,10 @@ class RelatedArticlesHandler implements
 	 *
 	 * @param Title $title
 	 * @param DBConnRef $dbr
+	 * @param array $targetNamespaces Empty array means accepting all namespaces.
 	 * @return string
 	 */
-	private function makeTitlesToRedirectsOfHereSQL( Title $title, DBConnRef $dbr ): string {
+	private function makeTitlesToRedirectsOfHereSQL( Title $title, DBConnRef $dbr, $targetNamespaces ): string {
 		$redirects = $dbr->newSelectQueryBuilder()
 			->table( 'redirect', 'redirect' )
 			->leftJoin( 'page', 'page', [
@@ -234,6 +246,9 @@ class RelatedArticlesHandler implements
 			->leftJoin( 'page', 'redirect_page', [
 				'page_id = pl_from',
 			] )
+			->conds( $targetNamespaces ? [
+				'redirect_page.page_namespace' => $targetNamespaces,
+			] : [] )
 			->fields( [
 				'page_namespace' => 'redirect_page.page_namespace',
 				'page_title' => 'redirect_page.page_title',